run tests on server version only if content update was successfull
