+++ .circleci/config.yml
@@ -314,8 +314,9 @@ jobs:
                 echo "Skipping - not running in INSTANCE_TESTS build, unit-tests container or NIghtly run"
                 exit 0
             fi
-            if ./Tests/scripts/is_ami.sh ;
+            if ./Tests/scripts/is_ami.sh;
               then
+                python ./Tests/scripts/is_content_installation_successful.py
                 ./Tests/scripts/run_tests.sh "Demisto GA"
 
               else
@@ -337,6 +338,7 @@ jobs:
             fi
             if ./Tests/scripts/is_ami.sh ;
               then
+                python ./Tests/scripts/is_content_installation_successful.py
                 ./Tests/scripts/run_tests.sh "Demisto one before GA"
 
               else
@@ -358,6 +360,7 @@ jobs:
             fi
             if ./Tests/scripts/is_ami.sh ;
               then
+                python ./Tests/scripts/is_content_installation_successful.py
                 ./Tests/scripts/run_tests.sh "Demisto two before GA"
 
               else
@@ -379,6 +382,7 @@ jobs:
             fi
             if ./Tests/scripts/is_ami.sh ;
               then
+                python ./Tests/scripts/is_content_installation_successful.py
                 ./Tests/scripts/run_tests.sh "Server Master"
                 export RETVAL=$?
                 cp ./Tests/failed_tests.txt $CIRCLE_ARTIFACTS/failed_tests.txt
